+++ client/src/components/pages/Dashboard.jsx
@@ -1,8 +1,10 @@
 import React from "react";
+import GoogleLogin from "../molecules/GoogleLogin";
 import Login from "../molecules/Login";
 const Dashboard = () => {
   return (
     <div>
+      <GoogleLogin />
       <Login />
     </div>
   );
